requirements.txt
streamlit==1.38.0
pandas==2.2.3
numpy==1.26.4
plotly==5.23.0
python-dateutil==2.9.0.post0
pytest==8.1.1
fpdf2==2.7.9
kaleido==0.2.1
